Relocating the discovery service.
